+++ lib/event_bloc_no_widgets.dart
@@ -4,5 +4,6 @@
 library event_bloc_no_widgets;
 
 export 'package:event_bloc/src/event_bloc/bloc.dart';
+export 'package:event_bloc/src/event_bloc/event.dart';
 export 'package:event_bloc/src/event_bloc/event_channel.dart';
 export 'package:event_bloc/src/event_bloc/repository.dart';