+++ lib/src/event_bloc/event.dart
@@ -1,0 +1,6 @@
+class BlocEvent<T> {
+  final String eventName;
+
+  const BlocEvent(this.eventName);
+  const BlocEvent.fromObject(Object eventType) : eventName = '$eventType';
+}